- Removed old code from Django manage.py. - Added a utility script for adding whitelist entries directly based on PEM certificates. - Added configuration for Django's new ALLOWED_HOSTS setting. - Switched to a certificate naming convention where the extension follows the regular convention of representing the file format, not the contents - Added configuration setting that allows disabling whitelist verification for MNStorage.update(). - Added code so that GMN automatically trusts its own client side certificate for incoming connections. This is safe and simplifies stand-alone deployment. - Removed code left over from when the GMN async processes authenticated via HTTP and not via the database. - Fixed bug that limited size of objects in GMN to 4GB. - Removed the internal API used by the old version of the async processes. - Removed concurrency tests that used to be necessary for checking that Apache was configured correctly. - Added ".conf" to Apache configuration files to match new naming requirement in 14.04. - Updated virtual host configuration file to use Apache's new security directives. - Switched to new configuration type in Apache that uses one or more WSGI Daemon Processes instead of directly hosting mod_wsgi. This allows GMN to run with the "gmn" user instead of the Apache user, which simplifies file permissions. It also makes it easier to configure desired number of processes and threads for GMN. - Removed custom apache2.conf file. It's no longer necessary since GMN now can run directly with any of Apache's MPMs. - Removed the custom Apache ports.conf file. It's no longer necessary because Apache has simplified setting up a Virtual Host. - Bumped version numbers for dependencies and for GMN itself.

+++ d1_mn_generic/src/service/mn/models.py
@@ -54,10 +54,10 @@ class ScienceObject(models.Model):
   checksum = models.CharField(max_length=128, db_index=True)
   checksum_algorithm = models.ForeignKey(ScienceObjectChecksumAlgorithm, db_index=True)
   mtime = models.DateTimeField(db_index=True)
-  size = models.PositiveIntegerField(db_index=True)
+  size = models.BigIntegerField(db_index=True)
   replica = models.BooleanField(db_index=True)
   system_metadata_refreshed = models.DateTimeField(null=True)
-  serial_version = models.BigIntegerField()
+  serial_version = models.PositiveIntegerField()
   archived = models.BooleanField()
 
   def set_format(self, format_id):
@@ -183,7 +183,7 @@ class SystemMetadataDirtyQueueStatus(models.Model):
 class SystemMetadataDirtyQueue(models.Model):
   object = models.ForeignKey(ScienceObject)
   timestamp = models.DateTimeField(auto_now=True)
-  serial_version = models.BigIntegerField()
+  serial_version = models.PositiveIntegerField()
   last_modified = models.DateTimeField()
   status = models.ForeignKey(SystemMetadataDirtyQueueStatus)
